+++ aaanalysis/__init__.py
@@ -8,7 +8,8 @@ from .feature_engineering import (AAclust, AAclustPlot, SequenceFeature, Numeric
                                   CPPGrid, CPPPlot, SequenceFeatureTransformer)
 from .pu_learning import dPULearn, dPULearnPlot
 from .explainable_ai import TreeModel
-from .prediction import AAPred, AAPredPlot, ReliabilityModel, ReliabilityModelPlot
+from .prediction import (AAPred, AAPredPlot, ReliabilityModel, ReliabilityModelPlot,
+                         ModelEvaluator, ModelEvaluatorPlot)
 from .protein_engineering import AAMut, AAMutPlot, SeqMut, SeqMutPlot, SeqOpt, SeqOptPlot
 from .plotting import (plot_get_clist, plot_get_cmap, plot_get_cdict,
                        plot_settings, plot_legend, plot_gcfs)
@@ -67,6 +68,8 @@ __all__ = [
     "AAPredPlot",
     "ReliabilityModel",
     "ReliabilityModelPlot",
+    "ModelEvaluator",
+    "ModelEvaluatorPlot",
     # "ShapModel"       # SHAP
     "plot_get_clist",
     "plot_get_cmap",